relocate configuration into core
